[context] Music: Publish/use focal Entities
In the Music mod, publish to the focal_entities topic. In the concerts
Agent, look for the story-level focal_entities and propose a concert
for every artist found there.

This change doesn't address Kronk (obviously). Kronk is currently broken
for other reasons, but will be updated once working at all to use the
same pattern as here.

Change-Id: I56001b957ebb2c814f30cfa8951a560cd70bdb13